requirements.txt
requests
beautifulsoup4
pandas
python-telegram-bot
python-decouple
fastapi
uvicorn
python-dotenv
websockets